Replace tasks services to new domain area
