+++ lib/tasks/interactors/issue_information.rb
@@ -28,8 +28,8 @@ module Tasks
 
       def match_host(issue_url)
         Container['matchers.git_host'].(issue_url) do |m|
-          m.success(:github) { |issue_data| Container['services.github_issue_requester'].(issue_data) }
-          m.success(:gitlab) { |issue_data| Container['services.gitlab_issue_requester'].(issue_data) }
+          m.success(:github) { |issue_data| Container['tasks.services.github_issue_requester'].(issue_data) }
+          m.success(:gitlab) { |issue_data| Container['tasks.services.gitlab_issue_requester'].(issue_data) }
           m.failure do
             error('invalid url')
             INVALID_URL_ERROR
